Add Shared<>, Unique<>, Timer<> and Timer1
This enables one shot and recurring timer callbacks using a reasonable
syntax.

I had to change the way that we codegen the flags for the various
registers because the labels were different between atmega32u4 and
atmega328p.  I like this version a bit better because the names are
more concise too.